requirements.txt
PyPDF2
llama-index
llama-index-llms-gemini
tqdm
google.generativeai
typer